+++ Chalkable.Data.School/Model/Infraction.cs
@@ -5,11 +5,11 @@ namespace Chalkable.Data.School.Model
 {
     public class Infraction
     {
-        public short Id { get; set; }
+        public int Id { get; set; }
         public string Name { get; set; }
         public string Code { get; set; }
 	    public string Description { get; set; }
-	    public int Demerits { get; set; }
+	    public byte Demerits { get; set; }
 	    public string StateCode { get; set; }
         public string SIFCode { get; set; }
         public string NCESCode { get; set; }
